+++ CHANGELOG.md
@@ -6,3 +6,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## Unreleased
+### Added
+- Added `Time` value object. It represents time in the ISO-8601 calendar system (without time-zone).